+++ src/killstreaker.rs
@@ -1,6 +1,7 @@
 use strum_macros::{Display, EnumString, EnumIter};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
+use crate::Attribute;
 
 #[derive(Serialize_repr, Deserialize_repr, Debug, Hash, Eq, PartialEq, Display, EnumString, EnumIter, TryFromPrimitive, IntoPrimitive, Clone)]
 #[repr(u32)]
@@ -19,4 +20,8 @@ pub enum Killstreaker {
     Incinerator = 2007,
     #[strum(serialize = "Hypno-Beam")]
     HypnoBeam = 2008,
+}
+
+impl Attribute for Killstreaker {
+    const DEFINDEX: u32 = 2013;
 }