+++ src/sheen.rs
@@ -1,6 +1,7 @@
 use strum_macros::{Display, EnumString, EnumIter};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
+use crate::Attribute;
 
 #[derive(Serialize_repr, Deserialize_repr, Debug, Hash, Eq, PartialEq, Display, EnumString, EnumIter, TryFromPrimitive, IntoPrimitive, Clone)]
 #[repr(u8)]
@@ -19,4 +20,8 @@ pub enum Sheen {
     VillainousViolet = 6,
     #[strum(serialize = "Hot Rod")]
     HotRod = 7,
+}
+
+impl Attribute for Sheen {
+    const DEFINDEX: u32 = 2014;
 }