+++ src/traits.rs
@@ -1,0 +1,7 @@
+pub trait Attribute {
+    const DEFINDEX: u32;
+}
+
+pub trait Attributes {
+    const DEFINDEX: &'static [u32];
+}